requirements.txt
torch-directml

--extra-index-url https://download.pytorch.org/whl/cu121
torch
torchvision

imageio
tensorboard
PyYAML